backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
langchain
chromadb
openai
tiktoken
langchain-text-splitters
python-multipart
chromadb
sentence-transformers
google-genai
groq